backend/azure_function/log_uploader/requirements.txt
azure-functions
azure-search-documents==11.7.0b1
langchain-openai
azure-monitor-opentelemetry
langsmith
langfuse
python-dotenv
